requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
Pillow==9.4.0
requests==2.28.2
python-dotenv==1.0.0
google-cloud-vision==3.4.0
gunicorn==20.1.0 